// src/enums/socketEvent.enum.ts
export enum SocketEvent {
    SERVER_CONNECTION = 'connection',
    DISCONNECT = 'disconnect',
    AUTHETICATION = 'authenticate',
    LOGIN = 'login',
    REQUEST  = 'request',
    RESPONSE = 'response',
    SERVER_MESSAGE = 'server:message',
    CLIENT_CONNECT = 'connect',
    CLIENT_DISCONNECT = 'client:disconnect',
    CLIENT_MESSAGE = 'client:message',
}